chore: relocate manifest download and handling functions
